Hack: Disable splitting AAS world in clusters
Current Warsow executables have bugs related
to clusters handling that lead to unreachability
of places on a map that are often important.

I remember fixing these Warsow bugs in August 2018
and thought the fix has been committed but
unfortunately it has been lost.

This approach is not even that bad as it sounds.
The lower path-finding performance should not be
a concern considering the fact the router code
has been substantially optimized and a 2-layer
cache is used on top of it. Besides that
clustering can produce weird routing results.
There was a comment "there might be a shorter route
via another cluster but we don't care".
If Warsow code is going to expect a single cluster
this would simplify it substantially. And also
automatically detected clusers are rather poor.